- removed ugly br - outsourced javascript to own file
